Jquery with Spring
Programms run with redirect.jsp.
Controller request mapped with "view"  calls "jsp/index" depending on prefix and suffix mentioned in spring-servlet.xml.
In index.jsp bootstrap file link and jquery script file is added using the JSTL taglib.

After user input the two numbers and hit the submit button. it goes to index.jsp as mentioned above.
Once the document is ready depending on the button click it calls the function inside the javascript.
Here inside the function it calls the another controller method which is mapped with "math" and takes the two number as input argument and  multiplies those and returns the result in string format.
This returned output is taken using "resp" in function argument and displayed on browser using the html.